// apps/mobile/src/services/tradeCore/order/typings.d.ts
declare namespace Order {
// ...
  type BgaOrderPageListItem = {
    /**
     * 订单方向
     */
    buySell?: TradePositionDirectionEnum
    /**
     * 平仓价格
     */
    closePrice?: number
    /**
     * 配置
     */
    conf: Symbol.SymbolConf
    /**
     * 创建时间
     */
    createTime?: string
    /**
     * 数据源code
     */
    dataSourceCode?: string
    /**
     * 数据源Symbol
     */
    dataSourceSymbol?: string
    /**
     * 手续费
     */
    handlingFees?: number
    /**
     * 主键
     */
    id: string
    /**
     * 图标
     */
    imgUrl?: string
    /**
     * 库存费
     */
    interestFees?: number
    /**
     * 杠杆倍数
     */
    leverageMultiple?: number
    /**
     * 保证金类型
     */
    marginType?: OrderMarginTypeEnum
    /**
     * 订单模式
     */
    mode?: API.OrderMode
    /**
     * 订单保证金
     */
    orderMargin?: number
    /**
     * 订单保证金汇率
     */
    marginExchangeRate?: string
    /**
     * 订单基础保证金
     */
    orderBaseMargin?: number
    /**
     * 订单数量
     */
    orderVolume?: number
    /**
     * 盈亏
     */
    profit?: number
    /**
     * 备注
     */
    remark?: string
    /**
     * 开仓价格
     */
    startPrice?: number
    /**
     * 状态
     */
    status?: TradePositionStatusEnum
    /**
     * 止损
     */
    stopLoss?: number
    /**
     * 交易品种
     */
    symbol?: string
    /**
     * 交易品种别名
     */
    alias?: string
    /**
     * 品种小数位
     */
    symbolDecimal?: number
    /**
     * 止盈
     */
    takeProfit?: number
    /**
     * 交易账户ID
     */
    tradeAccountId?: any
    /**
     * 更新时间
     */
    updateTime?: string
    /** 品种配置 */
    conf?: Symbol.SpreadConf
    /**
     * 账户id
     */
    accountId: string
    /**
     * 交易账户001
     */
    accountName: string
    /**
     * 用户登录账号 654321@163.com
     */
    userAccount: string
    /** 用户名称 */
    userName: string
    /** 追加预付款比例 */
    addAdvanceCharge: number
    /** 强制平仓比例 */
    compelCloseRatio: number
    /** 订单类型 */
    type?: OrderTypeEnum
    /** 限价价格 */
    limitPrice?: number
    /** 账户组ID */
    accountGroupId?: string
    /** 平仓时间 */
    finishTime?: string
  }
// ...
}